trail-microservice/src/requirements.txt
Flask
Flask-SQLAlchemy
Flask-RESTful
Flask-Swagger-UI
pyodbc
pydantic
python-dotenv
requests
swagger-ui-bundle

flasgger

pyjwt